day 2 first hour work completed

+++ utils/cloudinary.js
@@ -1,3 +1,4 @@
+require("dotenv").config();
 const cloudinary = require("cloudinary").v2;
 const fs = require("fs");
 cloudinary.config({
@@ -9,14 +10,17 @@ cloudinary.config({
 const uploadOnCloudinary = async (localfilePath) => {
   try {
     if (!localfilePath) return null;
+    // console.log("localfilePath : ", localfilePath);
     const res = await cloudinary.uploader.upload(localfilePath, {
       resource_type: "auto",
     });
-    // console.log("File uploaded : ", res.secure_url);
+
+    // console.log("response : ", res);
     fs.unlinkSync(localfilePath);
     return res.secure_url;
   } catch (error) {
-    fs.unlinkSync(localfilePath);
+    console.error("Cloudinary Upload Error:", error);
+    if (fs.existsSync(localfilePath)) fs.unlinkSync(localfilePath);
     return null;
   }
 };
